Fixed Linux bugs. For building used GCC 4.7.2, all relative bugs/warnings are fixed. Fixed thread safety of __malloc_hook, __realloc_hook, __realloc_hook. Fixed logging time initialization.
Improved performance of OpenGL rendering.
Client and mapper GUI back to work in one thread with logic.

Client and mapper now use Script::ScriptGarbager like server, other stuff is erased.

Improved client script function SetEffect.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@483 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 503fce0fd2d1802d8b598256d77dd5195e7c4436

+++ Source/ResourceManager.cpp
@@ -189,7 +189,7 @@ AnyFrames* ResourceManager::GetAnim( uint name_hash, int dir, int res_type )
 
 uint AnimMapId( uint crtype, uint anim1, uint anim2, int dir, bool is_fallout )
 {
-    uint dw[ 5 ] = { crtype, anim1, anim2, dir, is_fallout ? -1 : 1 };
+    uint dw[ 5 ] = { crtype, anim1, anim2, (uint) dir, is_fallout ? uint( -1 ) : 1 };
     return Crypt.Crc32( (uchar*) &dw[ 0 ], sizeof( dw ) );
 }
